+++ demo/App.vue
@@ -1,22 +1,14 @@
+<script setup lang="ts">
+import DemoComponent from './DemoComponent.vue'
+</script>
+
 <template>
   <div class="home-main">
     <div class="hero">
-      <Demo />
+      <DemoComponent />
       <footer>
         <i>&copy;{{ new Date().getFullYear() }} dbetka</i>
       </footer>
     </div>
   </div>
 </template>
-
-<script lang="ts">
-import Demo from './Demo.vue'
-import { defineComponent } from 'vue'
-
-export default defineComponent({
-  name: 'App',
-  components: {
-    Demo,
-  },
-})
-</script>