+++ src/MaterialSymbol.vue
@@ -2,7 +2,6 @@
 import { computed } from 'vue'
 import { IconsTypesProp } from '@/jscache/icons-types'
 import { IconsProp } from '@/jscache/icons-names'
-import { useIcons } from '@/index'
 
 const props = withDefaults(
   defineProps<{
@@ -14,15 +13,13 @@ const props = withDefaults(
     filled?: boolean,
   }>(),
   {
-    type: () => window.$materialIconsDefaults?.defaultType || 'outlined',
-    weight: () => window.$materialIconsDefaults?.defaultWeight || '300',
-    grade: () => window.$materialIconsDefaults?.defaultGrade || 'medium',
-    size: () => window.$materialIconsDefaults?.defaultSize || 24,
-    filled: () => window.$materialIconsDefaults?.defaultFilled || false,
+    type: () => window.$materialSymbolsDefaults?.defaultType || 'outlined',
+    weight: () => window.$materialSymbolsDefaults?.defaultWeight || '300',
+    grade: () => window.$materialSymbolsDefaults?.defaultGrade || 'medium',
+    size: () => window.$materialSymbolsDefaults?.defaultSize || 24,
+    filled: () => window.$materialSymbolsDefaults?.defaultFilled || false,
   }
 )
-
-const icons = useIcons()
 
 const GRADES = {
   thin: -50,
@@ -36,8 +33,6 @@ const styles = computed(() => {
   const grade = GRADES[props.grade]
   const size = props.size
 
-  console.log(size)
-
   return [
     `font-size: ${size}px;`,
     `font-variation-settings: 'FILL' ${fill}, 'wght' ${weight}, 'GRAD' ${grade}, 'opsz' ${size};`,
@@ -46,9 +41,9 @@ const styles = computed(() => {
 
 const classes = computed(() => {
   switch (props.type) {
-    case icons.types.outlined: return 'material-symbols-outlined'
-    case icons.types.rounded: return 'material-symbols-rounded'
-    case icons.types.sharp: return 'material-symbols-sharp'
+    case 'outlined': return 'material-symbols-outlined'
+    case 'rounded': return 'material-symbols-rounded'
+    case 'sharp': return 'material-symbols-sharp'
     default: return ''
   }
 })