+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.0] - 2023-12.09
+### Changed
+- Change component name and way to import component.
+- Update exported types.
+- Update readme file.
+- Update demo.
+
 ## [1.0.3] - 2023-12.01
 ### Changed
 - Update readme file and fix demo.